Put definitions of Bebop functions in a separate .cc file
