+++ bebop_test/Makefile
@@ -1,5 +1,5 @@
 CXXFLAGS+=-std=c++14 -march=native -I.. -Wall -Wpedantic
-LDFLAGS=-g -pthread -lstdc++ -lm
+LDFLAGS=-g -pthread -lstdc++ -lm ../libbebop/bebop.o
 
 # OpenCV
 CXXFLAGS+=`pkg-config --cflags opencv`
@@ -9,13 +9,16 @@ LDFLAGS+=`pkg-config --libs opencv`
 LDFLAGS+=-lavcodec -lavformat -lavutil -lswscale
 
 # AR SDK
+# We set the rpath so that compiled programs can find the folder with the ARSDK
+# libs in, but apparently setting the rpath is deprecated; I'm sure there is a
+# nicer way to do this.
 ifndef ARSDK_ROOT
 	$(error Environment variable ARSDK_ROOT must be defined)
 endif
 AR_STAGING_PATH=$(ARSDK_ROOT)/out/arsdk-native/staging
 AR_LIB_PATH=$(AR_STAGING_PATH)/usr/lib
 CXXFLAGS+=-I$(AR_STAGING_PATH)/usr/include \
-	-Wl,-rpath,$(AR_LIB_PATH),--disable-new-dtags -Wno-deprecated-declarations
+	-Wl,-rpath,$(AR_LIB_PATH),--disable-new-dtags
 LDFLAGS+=$(AR_LIB_PATH)/libarsal.so $(AR_LIB_PATH)/libardiscovery.so \
 	$(AR_LIB_PATH)/libarcontroller.so $(AR_LIB_PATH)/libarnetworkal.so \
 	$(AR_LIB_PATH)/libarcommands.so $(AR_LIB_PATH)/libmux.so \
@@ -25,14 +28,17 @@ LDFLAGS+=$(AR_LIB_PATH)/libarsal.so $(AR_LIB_PATH)/libardiscovery.so \
 	$(AR_LIB_PATH)/libsdp.so $(AR_LIB_PATH)/libulog.so \
 	$(AR_LIB_PATH)/libarmedia.so $(AR_LIB_PATH)/libfutils.so
 
-.PHONY: all clean
+.PHONY: all clean libbebop
 
 TARGET := bebop_test
 
 all: $(TARGET)
 
-$(TARGET): $(TARGET).cc ../robots/bebop.h
+$(TARGET): $(TARGET).cc libbebop
 	$(CXX) $(CXXFLAGS) -o $(TARGET) $(TARGET).cc $(LDFLAGS)
+
+libbebop:
+	$(MAKE) -C ../libbebop
 
 clean:
 	rm -f $(TARGET)
